feat[GLHL-40]: card fix with column and board

+++ src/main/java/com/sparta/trello/domain/card/entity/Card.java
@@ -38,7 +38,7 @@ public class Card extends BaseEntity {
     @JoinColumn(name = "trelloColumn_id",nullable = false)
     private TrelloColumn trelloColumn;
 
-    private int position;
+    private long position;
 
     @Builder.Default
     @Enumerated(EnumType.STRING)
